# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SubmitRequest": {
        "additionalProperties": false,
        "description": "The native submission body. ``max_new_tokens`` is capped at the documented ceiling (E-08, E-18).\n\n``extra=\"forbid\"`` rejects any non-native field (a cloud field such as ``model``/``prompt``/\n``audio_setting``, or ``stream``) with HTTP 422 rather than silently discarding it \u2014 cloud-schema\ntranslation is S4's, and S2 must not let an unsupported field appear honoured (INV-7 boundary).",
        "properties": {
          "input": {
            "minLength": 1,
            "title": "Input",
            "type": "string"
          },
          "instructions": {
            "minLength": 1,
            "title": "Instructions",
            "type": "string"
          },
          "max_new_tokens": {
            "maximum": 9000.0,
            "minimum": 1.0,
            "title": "Max New Tokens",
            "type": "integer"
          },
          "seed": {
            "default": 0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "input",
          "instructions",
          "max_new_tokens"
        ],
        "title": "SubmitRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MiniMax-Music3 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/artifacts/{job_id}/{name}": {
      "get": {
        "description": "Serve one canonical member of a completed artifact by streaming an O_NOFOLLOW-validated fd; 404 for\na crafted id, a non-member name, an incomplete directory, a symlinked member, or an unknown id.",
        "operationId": "get_artifact_artifacts__job_id___name__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact"
      }
    },
...
    "/jobs": {
      "post": {
        "description": "Accept a native submission, queue it for the single worker, and return the queued job (202).",
        "operationId": "submit_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Job Jobs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Job"
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "description": "Return the job's current status view, or 404 if the id is unknown.",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "description": "Cancel a queued or running job (200); 404 if unknown, 409 if already terminal.",
        "operationId": "cancel_job_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Job Jobs  Job Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job"
      }
    },
    "/jobs/{job_id}/events": {
      "get": {
        "description": "Stream the job's lifecycle events as SSE, honoring ``Last-Event-ID``; 404 if the id is unknown.",
        "operationId": "stream_events_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events"
      }
    },
...
  }
}